Fix: Corrected quaternion conversion to Unified Fix: Cleaned up code-base, moved XR specific functions to unique class. Change: Renamed default namespace.
