+++ VRCFT - Quest OpenXR/VRCFT - Quest OpenXR.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
-    <RootNamespace>VRCFT___Quest_OpenXR</RootNamespace>
+    <RootNamespace>Quest_OpenXR</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <Platforms>AnyCPU;x64</Platforms>
